Move ReactionsMenu preview/snapshot to sheet level
The previous content-level Sample (`ReactionsMenuContentOneReaction` /
`ManyReactions`) bypassed the public `ReactionsMenu` composable and the
StreamCardBottomSheet wrapper — they predate sheet-level rendering being
viable in @Preview and Paparazzi. Now that both surfaces render the
real sheet path, drive the previews and snapshots through
`ReactionsMenuSample*` instead, which calls the public `ReactionsMenu`
end-to-end. Closes the 0% new-code coverage gap on the file.

Renames `ReactionsMenuContentTest` → `ReactionsMenuTest` and relocates
it under `ui.components.messageactions` (the package of the file under
test). Switches from `snapshotWithDarkMode` to `snapshot` + light/dark
pairs, matching the pattern used by every other sheet test.
